partitionid use uint64

+++ proto/extent_key.go
@@ -30,7 +30,7 @@ var (
 
 type ExtentKey struct {
 	FileOffset   uint64
-	PartitionId  uint32
+	PartitionId  uint64
 	ExtentId     uint64
 	ExtentOffset uint64
 	Size         uint32
